+++ http/open.go
@@ -8,13 +8,13 @@ import (
 	"net/http"
 )
 
-//Download downloads asset for supplied object
-func (s *manager) Download(ctx context.Context, object storage.Object, options ...storage.Option) (io.ReadCloser, error) {
-	return s.DownloadWithURL(ctx, object.URL(), options...)
+//Open downloads asset for supplied object
+func (s *manager) Open(ctx context.Context, object storage.Object, options ...storage.Option) (io.ReadCloser, error) {
+	return s.OpenURL(ctx, object.URL(), options...)
 }
 
-//Download downloads asset for supplied object
-func (s *manager) DownloadWithURL(ctx context.Context, URL string, options ...storage.Option) (io.ReadCloser, error) {
+//Open downloads asset for supplied object
+func (s *manager) OpenURL(ctx context.Context, URL string, options ...storage.Option) (io.ReadCloser, error) {
 	request, err := http.NewRequest(http.MethodGet, URL, nil)
 	if err != nil {
 		return nil, err